move get_new_items_with_ids to session.py
