+++ etl-pipeline/api/assistant/session.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 from sqlalchemy import insert, select, update
 from sqlalchemy import text as sql_text
-
 from agents.extensions.memory import SQLAlchemySession
-from agents.items import TResponseInputItem
+from agents.items import TResponseInputItem, ToolApprovalItem
+from agents import RunResult
 
 
 class CustomSQLAlchemySession(SQLAlchemySession):
@@ -70,3 +70,38 @@ class CustomSQLAlchemySession(SQLAlchemySession):
     def inserted_items(self) -> list[tuple[int, TResponseInputItem]]:
         """All (db_id, item) pairs written by add_items() during this session object's lifetime."""
         return list(self._inserted_items)
+
+
+def get_new_items_with_ids(
+    run_result: RunResult,
+    session: CustomSQLAlchemySession,
+) -> list[dict]:
+    """
+    Returns items that are in both RunResult.new_items and CustomSQLAlchemySession.inserted_items (converted via .to_input_item()) , with DB `agent_messages.id` added as `db_id`.
+
+    If there are multiple identically valued items in .inserted_items the first
+    db_id for each match in .new_items is used (a pop-on-first-match pool correctly
+    handles duplicate item content.) In case of duplicates, if .add_items()
+    has been called outside of the run that produced RunResult, db_ids might
+    not be correct.
+
+    Note: ToolApprovalItem's are not persisted, thus are not in .inserted_items, and are not returned.
+
+    """
+    new_items = [
+        item.to_input_item()
+        for item in run_result.new_items
+        if not isinstance(item, ToolApprovalItem)
+    ]
+    # calling .to_input_item() on ToolApprovalItem raises (agents/items.py).
+    # And they are not persisted agents/run_internal/session_persistence.py, line 243.
+
+    messages = []
+    for db_id, inserted_item in session.inserted_items:
+        try:
+            idx = new_items.index(inserted_item)  # dict equality check
+            new_items.pop(idx)
+            messages.append({"db_id": db_id, **inserted_item})
+        except ValueError:
+            pass  # if `inserted_item` is not in `new_items` (.e.g. Runner.run(input=) items)
+    return messages
